removed extra files and renamed components to match files
