+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"/Users/riva/repos/todo/src/reportWebVitals.js":"1","/Users/riva/repos/todo/src/App.js":"2","/Users/riva/repos/todo/src/Components/FormComponent.js":"3","/Users/riva/repos/todo/src/index.js":"4"},{"size":362,"mtime":1606171492211,"results":"5","hashOfConfig":"6"},{"size":242,"mtime":1606174584908,"results":"7","hashOfConfig":"6"},{"size":1193,"mtime":1606174486934,"results":"8","hashOfConfig":"6"},{"size":336,"mtime":1606174416970,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"czx635",{"filePath":"12","messages":"13","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"14","messages":"15","errorCount":1,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/riva/repos/todo/src/reportWebVitals.js",[],"/Users/riva/repos/todo/src/App.js",["18"],"/Users/riva/repos/todo/src/Components/FormComponent.js",["19"],"/Users/riva/repos/todo/src/index.js",[],{"ruleId":"20","severity":1,"message":"21","line":2,"column":8,"nodeType":"22","messageId":"23","endLine":2,"endColumn":16},{"ruleId":null,"fatal":true,"severity":2,"message":"24","line":6,"column":22},"no-unused-vars","'TodoForm' is defined but never used.","Identifier","unusedVar","Parsing error: Unexpected token, expected \";\"\n\n  4 | const TodoForm = () => {\n  5 | \n> 6 |   constructor(props) {\n    |                      ^\n  7 |     super(props);\n  8 |     this.state = { item: {}};\n  9 |   }"]
