+++ src/Components/ToDoConnected.js
@@ -7,7 +7,7 @@ import './todo.scss';
 const todoAPI = 'https://api-js401.herokuapp.com/api/v1/todo';
 
 
-const ToDo = () => {
+const ToDoConnected = () => {
 
   const [list, setList] = useState([]);
 
@@ -89,4 +89,4 @@ const ToDo = () => {
   );
 };
 
-export default ToDo;
+export default ToDoConnected;
