+++ src/Components/TodoList.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-class TodoList extends React.Component {
+const TodoList = () => {
 
   render() {
     return (
